EditNoteCard component created and useTransition logic moved into component
